create SingleGlobalGenServer
